requirements.txt
streamlit
google-generativeai
pypdf
python-pptx
olefile
fpdf2
python-dotenv
pyyaml
